Update from master branch

+++ scripts/installCommonDeps.sh
@@ -1,4 +1,4 @@
-#!/bin/bash
+#!/bin/bash -e
 
 pause() {
   read -p "$*"
@@ -56,7 +56,17 @@ install_fdkaac(){
   local SRC_MD5SUM="13c04c5f4f13f4c7414c95d7fcdea50f"
 
   local LIST_LIBS=`ls ${PREFIX_DIR}/lib/libfdk* 2>/dev/null`
-  $INCR_INSTALL && [[ ! -z $LIST_LIBS ]] && echo "fdkaac already installed." && return 0
+  [ "$INCR_INSTALL" = true ] && [[ ! -z $LIST_LIBS ]] && \
+    echo "fdkaac already installed." && return 0
+
+  if [ "$CHECK_INSTALL" = true ]; then
+    if [[ ! -z $LIST_LIBS ]]; then
+      echo "libfdkaac - Yes"
+    else
+      echo "libfdkaac - No"
+    fi
+    return 0
+  fi
 
   mkdir -p ${LIB_DIR}
   pushd ${LIB_DIR}
@@ -75,14 +85,24 @@ install_fdkaac(){
 }
 
 install_ffmpeg(){
-  local VERSION="4.1.3"
+  local VERSION="4.4.1"
   local DIR="ffmpeg-${VERSION}"
   local SRC="${DIR}.tar.bz2"
   local SRC_URL="http://ffmpeg.org/releases/${SRC}"
-  local SRC_MD5SUM="9985185a8de3678e5b55b1c63276f8b5"
+  local SRC_MD5SUM="9c2ca54e7f353a861e57525ff6da335b"
 
   local LIST_LIBS=`ls ${PREFIX_DIR}/lib/libav* 2>/dev/null`
-  $INCR_INSTALL && [[ ! -z $LIST_LIBS ]] && echo "ffmpeg already installed." && return 0
+  [ "$INCR_INSTALL" = true ]  && [[ ! -z $LIST_LIBS ]] && \
+  echo "ffmpeg already installed." && return 0
+
+  if [ "$CHECK_INSTALL" = true ]; then
+    if [[ ! -z $LIST_LIBS ]]; then
+      echo "ffmpeg - Yes"
+    else
+      echo "ffmpeg - No"
+    fi
+    return 0
+  fi
 
   mkdir -p ${LIB_DIR}
   pushd ${LIB_DIR}
@@ -104,10 +124,20 @@ install_ffmpeg(){
 }
 
 install_zlib() {
-  local VERSION="1.2.11"
+  local VERSION="1.2.12"
 
   local LIST_LIBS=`ls ${PREFIX_DIR}/lib/libz* 2>/dev/null`
-  $INCR_INSTALL && [[ ! -z $LIST_LIBS ]] && echo "zlib already installed." && return 0
+  [ "$INCR_INSTALL" = true ]  && [[ ! -z $LIST_LIBS ]] && \
+  echo "zlib already installed." && return 0
+
+  if [ "$CHECK_INSTALL" = true ]; then
+    if [[ ! -z $LIST_LIBS ]]; then
+      echo "zlib - Yes"
+    else
+      echo "zlib - No"
+    fi
+    return 0
+  fi
 
   if [ -d $LIB_DIR ]; then
     pushd $LIB_DIR >/dev/null
@@ -129,7 +159,17 @@ install_zlib() {
 #libnice depends on zlib
 install_libnice0114(){
   local LIST_LIBS=`ls ${PREFIX_DIR}/lib/libnice* 2>/dev/null`
-  $INCR_INSTALL && [[ ! -z $LIST_LIBS ]] && echo "libnice already installed." && return 0
+  [ "$INCR_INSTALL" = true ] && [[ ! -z $LIST_LIBS ]] && \
+  echo "libnice already installed." && return 0
+
+  if [ "$CHECK_INSTALL" = true ]; then
+    if [[ ! -z $LIST_LIBS ]]; then
+      echo "libnice - Yes"
+    else
+      echo "libnice - No"
+    fi
+    return 0
+  fi
 
   if [ -d $LIB_DIR ]; then
     cd $LIB_DIR
@@ -151,7 +191,17 @@ install_libnice0114(){
 #libnice depends on zlib
 install_libnice014(){
   local LIST_LIBS=`ls ${PREFIX_DIR}/lib/libnice* 2>/dev/null`
-  $INCR_INSTALL && [[ ! -z $LIST_LIBS ]] && echo "libnice already installed." && return 0
+  [ "$INCR_INSTALL" = true ] && [[ ! -z $LIST_LIBS ]] && \
+  echo "libnice already installed." && return 0
+
+  if [ "$CHECK_INSTALL" = true ]; then
+    if [[ ! -z $LIST_LIBS ]]; then
+      echo "libnice - Yes"
+    else
+      echo "libnice - No"
+    fi
+    return 0
+  fi
 
   if [ -d $LIB_DIR ]; then
     cd $LIB_DIR
@@ -176,11 +226,21 @@ install_libnice014(){
 
 install_openssl(){
   local LIST_LIBS=`ls ${PREFIX_DIR}/lib/libssl* 2>/dev/null`
-  $INCR_INSTALL && [[ ! -z $LIST_LIBS ]] && echo "openssl already installed." && return 0
+  [ "$INCR_INSTALL" = true ] && [[ ! -z $LIST_LIBS ]] && \
+  echo "openssl already installed." && return 0
+
+  if [ "$CHECK_INSTALL" = true ]; then
+    if [[ ! -z $LIST_LIBS ]]; then
+      echo "openssl - Yes"
+    else
+      echo "openssl - No"
+    fi
+    return 0
+  fi
 
   if [ -d $LIB_DIR ]; then
     local SSL_BASE_VERSION="1.1.1"
-    local SSL_VERSION="1.1.1j"
+    local SSL_VERSION="1.1.1m"
     cd $LIB_DIR
     rm -f ./build/lib/libssl.*
     rm -f ./build/lib/libcrypto.*
@@ -203,7 +263,17 @@ install_openssl(){
 
 install_openh264(){
   local LIST_LIBS=`ls ${ROOT}/third_party/openh264/libopenh264* 2>/dev/null`
-  $INCR_INSTALL && [[ ! -z $LIST_LIBS ]] && echo "openh264 already installed." && return 0
+  [ "$INCR_INSTALL" = true ] && [[ ! -z $LIST_LIBS ]] && \
+  echo "openh264 already installed." && return 0
+
+  if [ "$CHECK_INSTALL" = true ]; then
+    if [[ ! -z $LIST_LIBS ]]; then
+      echo "openh264 - Yes"
+    else
+      echo "openh264 - No"
+    fi
+    return 0
+  fi
 
   MAJOR=1
   MINOR=7
@@ -244,11 +314,20 @@ install_openh264(){
 
 install_libexpat() {
   local LIST_LIBS=`ls ${PREFIX_DIR}/lib/libexpat* 2>/dev/null`
-  $INCR_INSTALL && [[ ! -z $LIST_LIBS ]] && echo "libexpat already installed." && return 0
+  if [ "$CHECK_INSTALL" = true ]; then
+    if [[ ! -z $LIST_LIBS ]]; then
+      echo "libexpat - Yes"
+    else
+      echo "libexpat - No"
+    fi
+    return 0
+  fi
+  [ "$INCR_INSTALL" = true ] && [[ ! -z $LIST_LIBS ]] && \
+  echo "libexpat already installed." && return 0
 
   if [ -d $LIB_DIR ]; then
-    local VERSION="2.2.6"
-    local DURL="https://github.com/libexpat/libexpat/releases/download/R_2_2_6/expat-${VERSION}.tar.bz2"
+    local VERSION="2.4.4"
+    local DURL="https://github.com/libexpat/libexpat/releases/download/R_2_4_4/expat-${VERSION}.tar.bz2"
     pushd ${LIB_DIR} >/dev/null
     rm -rf expat-*
     rm -f ./build/lib/libexpat.*
@@ -265,20 +344,40 @@ install_libexpat() {
   fi
 }
 
-install_webrtc79(){
-  $INCR_INSTALL &&  [[ -s $ROOT/third_party/webrtc-m79/libwebrtc.a ]] && \
+install_webrtc88(){
+  local OWT_WEBRTC_PATH="$ROOT/third_party/webrtc-m88"
+  if [ "$CHECK_INSTALL" = true ]; then
+    if [[ -s $OWT_WEBRTC_PATH/libwebrtc.a ]]; then
+      echo "webrtc88 - Yes"
+    else
+      echo "webrtc88 - No"
+    fi
+    return 0
+  fi
+
+  [ "$INCR_INSTALL" = true ] &&  [[ -s $OWT_WEBRTC_PATH/libwebrtc.a ]] && \
   echo "libwebrtc already installed." && return 0
 
-  [[ ! -d $ROOT/third_party/webrtc-m79 ]] && \
-    mkdir $ROOT/third_party/webrtc-m79
+  [[ ! -d $OWT_WEBRTC_PATH ]] && \
+    mkdir $OWT_WEBRTC_PATH
 
-  pushd ${ROOT}/third_party/webrtc-m79 >/dev/null
-  . $PATHNAME/installWebrtc.sh
-  popd
+  pushd $OWT_WEBRTC_PATH >/dev/null
+   . $PATHNAME/installWebrtc.sh
+  popd >/dev/null
 }
 
 install_webrtc(){
-  $INCR_INSTALL &&  [[ -s $ROOT/third_party/webrtc/libwebrtc.a ]] && \
+  install_webrtc88
+
+  if [ "$CHECK_INSTALL" = true ]; then
+    if [[ -s $ROOT/third_party/webrtc/libwebrtc.a ]]; then
+      echo "libwebrtc - Yes"
+    else
+      echo "libwebrtc - No"
+    fi
+    return 0
+  fi
+  [ "$INCR_INSTALL" = true ] &&  [[ -s $ROOT/third_party/webrtc/libwebrtc.a ]] && \
   echo "libwebrtc already installed." && return 0
 
   export GIT_SSL_NO_VERIFY=1
@@ -304,16 +403,19 @@ install_webrtc(){
   ./src/tools-woogeen/install.sh
   ./src/tools-woogeen/build.sh
   popd
-
-  install_webrtc79
-
-  if [[ "$OS" =~ .*ubuntu.* ]] && [[ "$OS_VERSION" =~ 20.04.* ]]; then
-    ${SUDO} update-alternatives --remove python /usr/bin/python2
-  fi
 }
 
 install_licode(){
-  $INCR_INSTALL && [[ -d ${ROOT}/third_party/licode ]] && echo "licode already installed." && return 0
+  if [ "$CHECK_INSTALL" = true ]; then
+    if [[ -d ${ROOT}/third_party/licode ]]; then
+      echo "licode - Yes"
+    else
+      echo "licode - No"
+    fi
+    return 0
+  fi
+  [ "$INCR_INSTALL" = true ] && [[ -d ${ROOT}/third_party/licode ]] && \
+  echo "licode already installed." && return 0
 
   local COMMIT="8b4692c88f1fc24dedad66b4f40b1f3d804b50ca" #v6
   local LINK_PATH="$ROOT/source/agent/webrtc/webrtcLib"
@@ -336,6 +438,18 @@ install_licode(){
 }
 
 install_quic(){
+  local LIST_LIBS=`ls ${PREFIX_DIR}/lib/libowt_web_transport.so 2>/dev/null`
+  if [ "$CHECK_INSTALL" = true ]; then
+    if [[ ! -z $LIST_LIBS ]]; then
+      echo "quic - Yes"
+    else
+      echo "quic - No"
+    fi
+    return 0
+  fi
+  [ "$INCR_INSTALL" = true ] && [[ ! -z $LIST_LIBS ]] && \
+  echo "quic already installed." && return 0
+
   # QUIC IO
   rm $ROOT/third_party/quic-lib -rf
   mkdir $ROOT/third_party/quic-lib
@@ -345,28 +459,30 @@ install_quic(){
   tar xzf dist.tgz
   popd
 
-  # QUIC transport
-  local QUIC_SDK_VERSION=`cat ${ROOT}/source/agent/addons/quic/quic_sdk_version`
-  local QUIC_TRANSPORT_PATH=${ROOT}/third_party/quic-transport
-  local QUIC_HEADERS_DIR=${ROOT}/build/libdeps/build/include/owt
-  if [ -d ${QUIC_TRANSPORT_PATH} ]; then
-    rm -r ${QUIC_TRANSPORT_PATH}
-  fi
-  mkdir ${QUIC_TRANSPORT_PATH}
-  pushd ${QUIC_TRANSPORT_PATH}
-  wget ${QUIC_TRANSPORT_PACKAGE_URL_PREFIX}/linux/${QUIC_SDK_VERSION}.zip
-  if [ $? -eq 0 ]; then
-    unzip ${QUIC_SDK_VERSION}.zip
-    rm ${QUIC_SDK_VERSION}.zip
-    cp bin/release/libowt_quic_transport.so ${ROOT}/build/libdeps/build/lib
-    if [ -d ${QUIC_HEADERS_DIR} ]; then
-      rm -r ${QUIC_HEADERS_DIR}
+  # WebTransport
+  if [ "$ENABLE_WEBTRANSPORT" == "true" ]; then
+    local QUIC_SDK_PACKAGE_NAME="quic-sdk-ubuntu-x64-ci"
+    local QUIC_SDK_URL=`cat ${ROOT}/source/agent/addons/quic/quic_sdk_url`
+    local QUIC_TRANSPORT_PATH=${ROOT}/third_party/quic-transport
+    local QUIC_HEADERS_DIR=${ROOT}/build/libdeps/build/include/owt
+    pushd ${QUIC_TRANSPORT_PATH}
+    if [ ! -d ${QUIC_HEADERS_DIR} ]; then
+      mkdir -p ${QUIC_HEADERS_DIR}
     fi
-    cp -r include/owt ${QUIC_HEADERS_DIR}
-  else
-    read -p "Failed to download prebuild QUIC SDK. Please download and compile QUIC SDK version ${QUIC_SDK_VERSION} from https://github.com/open-webrtc-toolkit/owt-deps-quic."
+    if [ ! -d ${ROOT}/build/libdeps/build/lib ]; then
+      mkdir -p ${ROOT}/build/libdeps/build/lib
+    fi
+    if [ -d bin/release ]; then
+      cp bin/release/libowt_web_transport.so ${ROOT}/build/libdeps/build/lib
+      if [ -d ${QUIC_HEADERS_DIR} ]; then
+        rm -r ${QUIC_HEADERS_DIR}
+      fi
+      cp -r include/owt ${QUIC_HEADERS_DIR}
+    else
+      read -p "Unable to find prebuild QUIC SDK. Please download it from ${QUIC_SDK_URL} with a GitHub personal access token or compile it from https://github.com/open-webrtc-toolkit/owt-sdk-quic."
+    fi
+    popd
   fi
-  popd
 }
 
 install_nicer(){
@@ -384,7 +500,16 @@ install_nicer(){
 
 install_libsrtp2(){
   local LIST_LIBS=`ls ${PREFIX_DIR}/lib/libsrtp2* 2>/dev/null`
-  $INCR_INSTALL && [[ ! -z $LIST_LIBS ]] && echo "libsrtp2 already installed." && return 0
+  if [ "$CHECK_INSTALL" = true ]; then
+    if [[ ! -z $LIST_LIBS ]]; then
+      echo "libsrtp2 - Yes"
+    else
+      echo "libsrtp2 - No"
+    fi
+    return 0
+  fi
+  [ "$INCR_INSTALL" = true ] && [[ ! -z $LIST_LIBS ]] && \
+  echo "libsrtp2 already installed." && return 0
 
   if [ -d $LIB_DIR ]; then
     cd $LIB_DIR
@@ -402,14 +527,22 @@ install_libsrtp2(){
 }
 
 install_node() {
-  local NODE_VERSION="v10.21.0"
-  echo -e "\x1b[32mInstalling nvm...\x1b[0m"
+  local NODE_VERSION="v14"
   NVM_DIR="${HOME}/.nvm"
 
-  $INCR_INSTALL && [[ -s "${NVM_DIR}/nvm.sh" ]] && \
+  if [ "$CHECK_INSTALL" = true ]; then
+    if [[ -s "${NVM_DIR}/nvm.sh" ]]; then
+      echo "node - Yes"
+    else
+      echo "node - No"
+    fi
+    return 0
+  fi
+  [ "$INCR_INSTALL" = true ] && [[ -s "${NVM_DIR}/nvm.sh" ]] && \
   echo "node already installed." && return 0
 
-  #install nvm
+  echo -e "\x1b[32mInstalling nvm...\x1b[0m"
+  # install nvm
   bash "${PATHNAME}/install_nvm.sh"
   #install node
   [[ -s "${NVM_DIR}/nvm.sh" ]] && . "${NVM_DIR}/nvm.sh"
@@ -419,6 +552,15 @@ install_node() {
 }
 
 install_node_tools() {
+  if [ "$CHECK_INSTALL" = true ]; then
+    npm list -g node-gyp > /dev/null 2>&1
+    if [ $? -eq 0 ]; then
+      echo "node_tools - Yes"
+    else
+      echo "node_tools - No"
+    fi
+    return 0
+  fi
   if [ "${INCR_INSTALL}" == "true" ]; then
     npm list -g node-gyp > /dev/null 2>&1
     [ $? -eq 0 ] && echo "node tools already installed." && return 0
@@ -427,7 +569,7 @@ install_node_tools() {
   check_proxy
   npm install -g --loglevel error node-gyp@6.1.0 grunt-cli underscore jsdoc
   pushd ${ROOT} >/dev/null
-  npm install nan@2.11.1
+  npm install nan@2.15.0
   pushd ${ROOT}/node_modules/nan >/dev/null
   patch -p1 < $PATHNAME/patches/nan.patch
   popd >/dev/null
@@ -437,7 +579,16 @@ install_node_tools() {
 # libre depends on openssl
 install_libre() {
   local LIST_LIBS=`ls ${PREFIX_DIR}/lib/libre* 2>/dev/null`
-  $INCR_INSTALL && [[ ! -z $LIST_LIBS ]] && echo "libre already installed." && return 0
+  if [ "$CHECK_INSTALL" = true ]; then
+    if [[ ! -z $LIST_LIBS ]]; then
+      echo "libre - Yes"
+    else
+      echo "libre - No"
+    fi
+    return 0
+  fi
+  [ "$INCR_INSTALL" = true ] && [[ ! -z $LIST_LIBS ]] && \
+  echo "libre already installed." && return 0
 
   if [ -d $LIB_DIR ]; then
     pushd ${LIB_DIR} >/dev/null
@@ -457,7 +608,16 @@ install_libre() {
 
 install_usrsctp() {
   local LIST_LIBS=`ls ${PREFIX_DIR}/lib/libusrsctp* 2>/dev/null`
-  $INCR_INSTALL && [[ ! -z $LIST_LIBS ]] && echo "usrsctp already installed." && return 0
+  if [ "$CHECK_INSTALL" = true ]; then
+    if [[ ! -z $LIST_LIBS ]]; then
+      echo "usrsctp - Yes"
+    else
+      echo "usrsctp - No"
+    fi
+    return 0
+  fi
+  [ "$INCR_INSTALL" = true ] && [[ ! -z $LIST_LIBS ]] && \
+  echo "usrsctp already installed." && return 0
 
   if [ -d $LIB_DIR ]; then
     local USRSCTP_VERSION="30d7f1bd0b58499e1e1f2415e84d76d951665dc8"
@@ -525,6 +685,18 @@ install_gcc(){
 }
 
 install_json_hpp() {
+  local LIST_LIBS=`ls ${PREFIX_DIR}/include/json.hpp 2>/dev/null`
+  if [ "$CHECK_INSTALL" = true ]; then
+    if [[ ! -z $LIST_LIBS ]]; then
+      echo "json_hpp - Yes"
+    else
+      echo "json_hpp - No"
+    fi
+    return 0
+  fi
+  [ "$INCR_INSTALL" = true ] && [[ ! -z $LIST_LIBS ]] && \
+  echo "json_hpp already installed." && return 0
+
   if [ -d $LIB_DIR ]; then
     local DOWNLOAD_JSON_LINK="https://github.com/nlohmann/json/releases/download/v3.6.1/json.hpp"
     pushd $LIB_DIR >/dev/null
@@ -539,27 +711,44 @@ install_json_hpp() {
 }
 
 install_svt_hevc(){
-    pushd $ROOT/third_party >/dev/null
-    rm -rf SVT-HEVC
-    git clone https://github.com/intel/SVT-HEVC.git
+  local LIST_LIBS=`ls ${PREFIX_DIR}/lib/libSvtHevcEnc* 2>/dev/null`
+  if [ "$CHECK_INSTALL" = true ]; then
+    if [[ ! -z $LIST_LIBS ]]; then
+      echo "svt_hevc - Yes"
+    else
+      echo "svt_hevc - No"
+    fi
+    return 0
+  fi
+  [ "$INCR_INSTALL" = true ] && [[ ! -z $LIST_LIBS ]] && \
+  echo "svt_hevc already installed." && return 0
 
-    pushd SVT-HEVC >/dev/null
-    git checkout v1.3.0
+  pushd $ROOT/third_party >/dev/null
+  rm -rf SVT-HEVC
+  git clone https://github.com/intel/SVT-HEVC.git
 
-    mkdir build
-    pushd build >/dev/null
-    cmake -DCMAKE_C_FLAGS="-std=gnu99" -DCMAKE_INSTALL_PREFIX=${PREFIX_DIR} ..
-    make && make install
-    popd >/dev/null
+  pushd SVT-HEVC >/dev/null
+  git checkout v1.3.0
 
-    # pseudo lib
-    echo \
-        'const char* stub() {return "this is a stub lib";}' \
-        > pseudo-svtHevcEnc.cpp
-    gcc pseudo-svtHevcEnc.cpp -fPIC -shared -o pseudo-svtHevcEnc.so
+  if [[ "$OS" =~ .*centos.* ]]
+  then
+    source /opt/rh/devtoolset-7/enable
+  fi
 
-    popd >/dev/null
-    popd >/dev/null
+  mkdir -p build
+  pushd build >/dev/null
+  cmake -DCMAKE_C_FLAGS="-std=gnu99" -DCMAKE_INSTALL_PREFIX=${PREFIX_DIR} ..
+  make && make install
+  popd >/dev/null
+
+  # pseudo lib
+  echo \
+      'const char* stub() {return "this is a stub lib";}' \
+      > pseudo-svtHevcEnc.cpp
+  gcc pseudo-svtHevcEnc.cpp -fPIC -shared -o pseudo-svtHevcEnc.so
+
+  popd >/dev/null
+  popd >/dev/null
 }
 
 cleanup_common(){
@@ -576,6 +765,18 @@ cleanup_common(){
 }
 
 install_boost(){
+  local LIST_LIBS=`ls ${PREFIX_DIR}/lib/libboost* 2>/dev/null`
+  if [ "$CHECK_INSTALL" = true ]; then
+    if [[ ! -z $LIST_LIBS ]]; then
+      echo "boost - Yes"
+    else
+      echo "boost - No"
+    fi
+    return 0
+  fi
+  [ "$INCR_INSTALL" = true ] && [[ ! -z $LIST_LIBS ]] && \
+  echo "boost already installed." && return 0
+
   if [ -d $LIB_DIR ]; then
     cd $LIB_DIR
     wget -c http://iweb.dl.sourceforge.net/project/boost/boost/1.65.0/boost_1_65_0.tar.bz2
